optimisation: remove Ticket class
this is a first step towards reducing memory use in the
counter, and reducing the use of classes which will be
instantiated (literally) millions of times.

note that the new TicketState class combines together
a ticket, and the preference it is up to. all logic
relating to a vote exhausting due to two preferences
at a given number (eg. two 3s) has been removed; when
loading the counter with data, it is the responsibility
of the caller to reduce the flows loaded in TicketState
to the minimum flow before the ticket exhausts.